Reworked Briefing module
