--- dzn_tSFramework/Modules/Briefing/Settings.sqf
@@ -1,5 +1,0 @@
-
-tSF_Briefing_ShowRoster		= true;
-tSF_Briefing_RosterTimeout		= -1;
-
-tSF_Briefing_UpdateRosterOnJIP 	= false;
