+++ dzn_tSFramework/Modules/Briefing/data/PreInit.sqf
@@ -1,0 +1,7 @@
+#include "script_component.hpp"
+
+__CLIENT_ONLY__
+
+LOG("Pre-Initialization started");
+
+INIT_COMPONENT;
