+++ dzn_tSFramework/Modules/Briefing/data/helper/bhelper.css
@@ -1,4 +1,4 @@
-body {	
+body {
 	margin: 0px;
 	padding: 0px;
 	font-family: helvetica,arial,sans-serif;
@@ -6,7 +6,7 @@ body {
 }
 
 /* ****************
-	Landing page 
+	Landing page
 **************** */
 #header {
 	background-color: #252525;
@@ -14,7 +14,7 @@ body {
 }
 
 #header-title {
-	font-size: 24px;	
+	font-size: 24px;
 	padding: 2px 0px 0px 40px;
 }
 
@@ -29,7 +29,7 @@ body {
 	font-size: 18px;
 	background-color: #5a5a5a;
 	padding: 6px 0px 3px 40px;
-	
+
 }
 
 /* ****************
@@ -64,20 +64,25 @@ body {
 	background-color: #CCCCCC;
 	font-size: 10px;
 	font-family: monospace;
-	overflow-y: scroll; 
+	/* overflow-y: scroll; */
 	height: 300px;
 	white-space:pre;
 }
 
+#result-tab-data textarea {
+    height: 95%;
+    width: 100%;
+}
+
 /* ****************
-	Form 
+	Form
 **************** */
 #briefing-form {
-	
+
 }
 .dl-1 {
 	display: inline-block;
-	width: 150px;	
+	width: 150px;
 	font-size: 15px;
 	font-size: 12px;
 	color: gray;
@@ -92,7 +97,7 @@ body {
 	width: 350px;
 	height: 20px;
 	padding: 3px 6px;
-	font-family: monospace;	
+	font-family: monospace;
 }
 
 .dl-3 {
@@ -113,7 +118,7 @@ body {
 }
 
 label {
-	display: inline-block;	
+	display: inline-block;
 	font-size: 15px;
 	font-size: 12px;
 	color: gray;
