+++ dzn_tSFramework/Modules/Briefing/data/script_component.hpp
@@ -1,0 +1,20 @@
+#include "..\..\script_macro.hpp"
+
+#define COMPONENT Briefing
+
+#define ROSTER_TOPIC_NAME "ORBAT"
+
+
+#define MAIN_WRAPPER "<font size='14' face='TahomaB'>%1</font>"
+#define HEADER_WRAPPER "<font color='%1'>%2</font> (%3)"
+#define ROLE_WRAPPER "<font color='#c9bea7'>%1</font>"
+#define NAME_WRAPPER "          <font size='12'>%1</font>"
+#define GEAR_WRAPPER "<font size='10' color='#cfcfcf'> / %1</font>"
+#define LR_RADIO_TAG " <font color='#93afba'>ДВ</font>"
+#define SW_RADIO_TAG " <font color='#9db099'>КВ</font>"
+
+#define SIDE_BLUFOR "BLUFOR"
+#define SIDE_OPFOR  "OPFOR"
+
+#define COLOR_DEFAULT "#ffffff"
+#define COLOR_GROUP_NAME "#f5c242"
