+++ dzn_tSFramework/Modules/Briefing/tSF_briefing.sqf
@@ -1,55 +1,84 @@
 //     tSF Briefing
 // Do not modify this part
-#define BRIEFING		_briefing = [];
-#define TOPIC(NAME) 	_briefing pushBack ["Diary", [ NAME,
-#define END			]];
-#define ADD_TOPICS	for "_i" from (count _briefing) to 0 step -1 do {player createDiaryRecord (_briefing select _i);};
-#define TAGS tSF_MissionTags =
+#define BRIEFING    _briefing = [];
+#define TOPIC(NAME) _briefing pushBack ["Diary", [ NAME,
+#define END         ]];
+#define ADD_TOPICS  for "_i" from (count _briefing) to 0 step -1 do {player createDiaryRecord (_briefing select _i);};
+#define TAGS        tSF_MissionTags =
 //
 //
 // Mission tags
-TAGS(["SPECOPS","INFANTRY","JTAC/CAS","RolePlay"]);
+TAGS(["INFANTRY"]);
 
 // Briefing goes here
 
 BRIEFING
-
 TOPIC("I. Обстановка:")
-"Описание ситуации"
+"### Краткое общее описание происходящего - когда, где, кто и почему участвует в миссии. Общие мотивы сторон. Т.к. мы играем каждый раз в разных временных периодах, за разные стороны, то обстановка должна давать понять кто мы, что происходит и как мы до такого дошли."
 END
 
 TOPIC("А. Враждебные силы:")
-"Описание вражеских сил"
+"### Краткое описание известных сил противника (состав и количество), например:
+<br />
+<br />1 взвод легкой пехоты.
+<br />
+<br />Также указывать известную или предпологаемую поддержку противника (бронетехника, минометы, гранатометы), описывать известные и ожидаемые места дислокации."
 END
 
 TOPIC("Б. Дружественные силы:")
-"Описание дружественных сил"
+"### Краткое описание сил игроков - 2 отделения Армии США: 1'1, 1'2. Детальное описание приветствуется. Например:
+<br />
+<br />1ый взвод 3ей роты 435ой парашютно-десантной дивизии тяжелых плуеметов:
+<br />
+<br />1`1 - отделение десантников (9 чел.),
+<br />1`4 - отделение поддержки (7 чел., 2 пулемета, 1 ПТРК)"
 END
 
 TOPIC("II. Задание:")
-"Описание задач"
+"1.
+<br />2.
+<br />3.
+<br />### Основные и дополнительные задачи которые должны быть выполнены в ходе миссии. Опциональные задачи должны быть указаны как “(опционально) Не уйти в запой”.
+<br />
+<br />Не забывайте - задачи это 1 возрождение для мертвых игроков - чем больше миссий, тем больше в ней должно быть задач!"
 END
 
 TOPIC("III. Выполнение:")
-"Описание указаний по выполнению"
+"### Описание основного предполагаемого плана выполнения (как задумано штабом и мишн мейкером)! Любые замечания по выполнению (план на уровне роты, предварительный план, правила обращения с гражданскими и т.п.), рекомендации по порядку выполнения задач, указание на фичи миссии (например, как отличить информатора от простого бандита). И только если совсем нечего сказать - “По плану командира”"
 END
 
 TOPIC("IV. Поддержка:")
-"Доступная поддержка"
+"1x Название_техники
+<br />3x Название_техники
+<br />
+<br />CCP
+<br />FARP
+<br />
+<br />Резерв (редеплой) - 1 на отряд
+<br />
+<br />### Описать доступные средства поддержки - транспорт (в формате - 1х автомобиль УАЗ (невооруженный), 2х бронеавтомобиль “Тигр” (вооруженный, АГС и ПКМ)”), артиллерия/авиация (с указанием позывных, если используется модуль из tSF), дополнительные боеприпасы/статики (указать где находятся), наличие CCP и FARP. Доступность и размер резервов для доукомплектации/развертывания."
 END
 
 TOPIC("V. Сигналы:")
-"Сигналы и радива"
+"PL NET 50 Mhz
+<br />SUP NET 51 Mhz
+<br />
+<br />1'1 - SR CH 1
+<br />1'2 - SR CH 2
+<br />1'3 - SR CH 3"
 END
 
 TOPIC("VI. Замечания:")
-"Прочие замечания"
+"### Дополнительные замечания, которые могут пригодится для понимания того как миссию выполнить правильно (например, краткая информация об использовании парашютов), какие-то дополнительные сведения об экипировке (например, кто несет заряды взрывчатки)"
 END
 
 if ((serverCommandAvailable '#logout') || !(isMultiplayer) || isServer) then {
-TOPIC("VII. GSO notes:")
-"-"
+TOPIC("VII. Замечания для GSO:")
+"### Информация которую увидит только ГСО. Указать какой тактики будет придерживаться враг, какие резервы он имеет и прочие советы по курированию миссии."
 END
 };
 
 ADD_TOPICS
+
+
+diag_log "BRIEFING ADDED";
